# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TACHIKOMA-OS Backend API",
    "description": "Central orchestrator API for the TACHIKOMA-OS AI ecosystem. Provides LLM gateway, GraphRAG memory, chat, agent tools, and data layer for microservices.",
    "contact": {
      "name": "TACHIKOMA-OS Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/llm/chat": {
      "post": {
        "tags": [
          "LLM Gateway"
        ],
        "summary": "POST /api/llm/chat",
        "description": "Non-streaming chat completion with message history",
        "operationId": "llm_chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chat completion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Chat error"
          }
        }
      }
    },
    "/api/llm/embed": {
      "post": {
        "tags": [
          "LLM Gateway"
        ],
        "summary": "POST /api/llm/embed",
        "description": "Generate embedding for a single text",
        "operationId": "llm_embed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embedding generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Embedding error"
          }
        }
      }
    },
    "/api/llm/generate": {
      "post": {
        "tags": [
          "LLM Gateway"
        ],
        "summary": "POST /api/llm/generate",
        "description": "Raw text generation from a prompt",
        "operationId": "llm_generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Text generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Generation error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "type": "object",
        "description": "=============================================================================\nChatMessage - Message in a conversation\n=============================================================================",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "model": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "EmbedRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "EmbedResponse": {
        "type": "object",
        "required": [
          "embedding",
          "dimensions"
        ],
        "properties": {
          "dimensions": {
            "type": "integer",
            "minimum": 0
          },
          "embedding": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          }
        }
      },
...
      "GenerateRequest": {
        "type": "object",
        "required": [
          "prompt"
        ],
        "properties": {
          "model": {
            "type": "string",
            "nullable": true
          },
          "prompt": {
            "type": "string"
          }
        }
      },
      "GenerateResponse": {
        "type": "object",
        "required": [
          "content",
          "model",
          "prompt_tokens",
          "completion_tokens"
        ],
        "properties": {
          "completion_tokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "content": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "prompt_tokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}